# feat: run as subcommand to support imports