--- a/BUCK
+++ b/BUCK
@@ -84,7 +84,6 @@ pnpm_workspace(
         "//app/auth-portal:package.json",
         "//app/web:package.json",
         "//bin/auth-api:package.json",
-        "//bin/lang-js:package.json",
         "//lib/eslint-config:package.json",
         "//lib/ts-lib:package.json",
         "//lib/tsconfig:package.json",