--- a/bin/lang-js/src/sandbox.ts
+++ b/bin/lang-js/src/sandbox.ts
@@ -3,16 +3,14 @@ import os from "node:os";
 import fs from "node:fs";
 import path from "node:path";
 import zlib from "node:zlib";
-import fetch from "node-fetch";
-import toml from "toml";
+
+import toml from "npm:toml";
 import jsonpatch from "npm:fast-json-patch";
-
-import * as _ from "npm:lodash-es";
+import * as _ from "https://deno.land/x/lodash_es@v0.0.2/mod.ts";
 import * as yaml from "npm:js-yaml";
-
 import Joi from "npm:joi";
+
 import { FunctionKind } from "./function.ts";
-import { makeConsole } from "./sandbox/console.ts";
 import { makeExec } from "./sandbox/exec.ts";
 import layout from "./sandbox/layout.ts";
 import template from "./sandbox/template.ts";
@@ -21,18 +19,15 @@ import {
   makeBeforeRequestStorage,
   makeMainRequestStorage,
 } from "./sandbox/requestStorage.ts";
-import { makePackage } from "./sandbox/package.ts";
 
 export type Sandbox = Record<string, unknown>;
 
 function commonSandbox(executionId: string): Sandbox {
   return {
-    console: makeConsole(executionId),
     _,
     Buffer,
     requestStorage: makeMainRequestStorage(),
     zlib,
-    fetch,
     siExec: makeExec(executionId),
     YAML: { stringify: yaml.dump, parse: yaml.load },
     os,
@@ -43,7 +38,6 @@ function commonSandbox(executionId: string): Sandbox {
     jsonpatch,
     layout,
     template,
-    pkg: makePackage(executionId),
   };
 }